exports only public funcs, fixes warnings

+++ BreezeEngine/src/ECS/Systems/CollisionSystem.cpp
@@ -9,7 +9,7 @@ void CollisionSystem::Update(ECSManager& ecs)
 
             // Sync position and size from RenderComponent to CollisionComponent
             const SDL_Rect& dstRect = ecs.renderComponents[e1].dstRect;
-            ecs.collisionComponents[e1].aabb = { dstRect.x, dstRect.y, dstRect.w, dstRect.h };
+            ecs.collisionComponents[e1].aabb = { (float)dstRect.x, (float)dstRect.y, (float)dstRect.w, (float)dstRect.h };
 
             for (int e2 = e1 + 1; e2 <= max_entity; e2++)
             {
@@ -18,7 +18,12 @@ void CollisionSystem::Update(ECSManager& ecs)
 
                     // Sync position and size from RenderComponent to CollisionComponent
                     const SDL_Rect& dstRect2 = ecs.renderComponents[e2].dstRect;
-                    ecs.collisionComponents[e2].aabb = { dstRect2.x, dstRect2.y, dstRect2.w, dstRect2.h };
+                    ecs.collisionComponents[e2].aabb = { 
+                        static_cast<float>(dstRect2.x),
+                        static_cast<float>(dstRect2.y), 
+                        static_cast<float>(dstRect2.w), 
+                        static_cast<float>(dstRect2.h)
+                    };
 
                     const AABB& box1 = ecs.collisionComponents[e1].aabb;
                     const AABB& box2 = ecs.collisionComponents[e2].aabb;
